[transfer-engine] Avoid excessive notification action updates. JB#51671
When there are rapid consecutive action name changes (due to the auto-
generated action name being different each time a progress notification
is created), this confuses the UI as it needs to rebuild the action
data, and by the time the user triggers an action, it may already have
changed the UI to refer to a different action underneath. To avoid
this, just use the same action name each time.