[transfer-engine] Introduce sharing plugin api v2. JB#62100
The existing sharing plugin mechanism wants to request all the methods
a plugin has support for. However on some cases it's better to ask
what methods there are for specific mime types, possibly with the
information whether it's for a single item or multiple ones.

Here is introduced a new version of the plugin api while still
keeping support for the old.

The SharingPluginLoader had some API changes. It doesn't anymore
request automatically the sharing methodsbut requires the application
using it to trigger a query with some hints on what's being shared.
Also some naming was adjusted as more precise. E.g. plugins() didn't
return a list of plugins but rather sharing methods. A plugin being
able to provide multiple methods.